Rename inputs.json file to avoid issues when submitting via AGC
When submitting to AGC, the AGC looks in the same dir as main.wdl
and fills inputs using files named inputs.*.json. It then overrides
these values using anything supplied in --inputsFile, but if there
are optional values that you do not define in this file they will be
picked up from the inputs.json file. In this case, this is a template
file that shows the appropriate types for values, so the workflow will
fail.