Only merge bams when a sample has > 1 input bam
Otherwise we will use the output of Whatshap haplotag for downstream
steps. This eliminates the need to generate a virtually identical
bam via copying or moving the file.
Updated whatshap haplotag to allow the user to set the output bam
name so that the output of merge_bams or whatshap haplotag when the
number of bams == 1 is the same.